Исправил схему БД, добавил таблицу со списками, которые будут добавлять себе пользователи, сделал половину функционала плейлистов

+++ gaming_progression_api/services/game_statuses.py
@@ -12,8 +12,14 @@ settings = get_settings()
 class StatusesService:
     async def change_status(self, uow: IUnitOfWork, new_status: ChangeGameStatus, user_id: UUID4):
         async with uow:
+            activity_id = await uow.activity_types.find_one(name=new_status.activity_type)
+            if not activity_id:
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail=f'Activity type not found',
+                )
             unique_string = await uow.statuses.find_one(
-                user_id=user_id, game_id=new_status.game_id, activity_id=new_status.activity_id,
+                user_id=user_id, game_id=new_status.game_id, activity_id=activity_id.id,
             )
             if unique_string:
                 await uow.statuses.delete_one(user_id=user_id, game_id=new_status.game_id)
@@ -25,14 +31,17 @@ class StatusesService:
 
             change_data = await uow.statuses.find_one(user_id=user_id, game_id=new_status.game_id)
             if change_data:
-                await uow.statuses.edit_one(data={'activity_id': new_status.activity_id})
+                await uow.statuses.edit_one(data={'activity_id': activity_id.id}, user_id=user_id, game_id=new_status.game_id)
                 await uow.commit()
                 raise HTTPException(
                     status_code=status.HTTP_200_OK,
-                    detail=f'Game {new_status.game_id} status changed to {new_status.activity_id}',
+                    detail=f'Game {new_status.game_id} status changed to {activity_id.id}',
                 )
             new_status = new_status.model_dump()
+            del new_status["activity_type"]
             new_status['user_id'] = user_id
+            new_status['activity_id'] = activity_id.id
+
             await uow.statuses.add_one(new_status)
             await uow.commit()
             raise HTTPException(
